refactor to allow import in consequences-api
