+++ paireddata/paireddata.go
@@ -1,33 +1,40 @@
 package paireddata
-import(
+
+import (
 	"sort"
 )
-type PairedData struct{
+
+// PairedData is paired data
+type PairedData struct {
 	Xvals []float64
 	Yvals []float64
 }
-type ValueSampler interface{
+
+// ValueSampler interface
+type ValueSampler interface {
 	SampleValue(inputValue interface{}) float64
 }
-func (p PairedData) SampleValue(inputValue interface{}) float64{
+
+// SampleValue implements ValueSampler
+func (p PairedData) SampleValue(inputValue interface{}) float64 {
 	xval, ok := inputValue.(float64)
-	if !ok{
+	if !ok {
 		return 0.0
 	}
-	if xval < p.Xvals[0]{
+	if xval < p.Xvals[0] {
 		return 0.0 //xval is less than lowest x value
 	}
 	size := len(p.Xvals)
-	if xval >= p.Xvals[size-1]{
+	if xval >= p.Xvals[size-1] {
 		return p.Yvals[size-1] //xval yeilds largest y value
 	}
-	if xval == p.Xvals[0]{
+	if xval == p.Xvals[0] {
 		return p.Yvals[0]
 	}
-	upper := sort.SearchFloat64s(p.Xvals,xval)
+	upper := sort.SearchFloat64s(p.Xvals, xval)
 	//interpolate
 	lower := upper - 1 // safe because we trapped the 0 case earlier
-	slope := (p.Yvals[upper] - p.Yvals[lower])/(p.Xvals[upper] - p.Xvals[lower])
+	slope := (p.Yvals[upper] - p.Yvals[lower]) / (p.Xvals[upper] - p.Xvals[lower])
 	a := p.Yvals[lower]
-	return a + slope* (xval-p.Xvals[lower])
+	return a + slope*(xval-p.Xvals[lower])
 }
